pyproject.toml
[project]
name = "prostate-cancer"
version = "0.1.0"
description = ""
authors = []
requires-python = ">=3.11,<3.12"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "numpy>=1.26.2",
    "pandas>=2.1.4",
    "scikit-learn>=1.3.2",
    "scipy>=1.11.4",
    "pyvips>=2.2.1",
    "torch>=2.1.2",
    "torchvision>=0.16.2",
    "torchmetrics>=1.2.1",
    "albumentations>=1.3.1",
    "lightning>=2.1.3",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
    "mlflow>=2.9.2",
    "humanize>=4.9.0",
    "scikit-image>=0.22.0",
]

[tool.pdm.dev-dependencies]
dev = ["mypy", "pre-commit", "ruff"]


[tool.pdm.scripts]
train = "python -m prostate_cancer.main stage=fit"
eval = "python -m prostate_cancer.mani stage=test"

l = { composite = ["lint", "format", "mypy"] }
lint = "ruff check"
format = "ruff format"
mypy = "mypy ."
post_install = { composite = [
    "pre-commit install",
    "pre-commit install --hook-type commit-msg",
] }

[tool.pdm]
distribution = false
